+++ delphi/latents/latents.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass, field
-from typing import Optional
+from typing import NamedTuple, Optional
 
 import blobfile as bf
 import orjson
@@ -7,6 +7,54 @@ from jaxtyping import Float
 from torch import Tensor
 from transformers import PreTrainedTokenizer, PreTrainedTokenizerFast
 
+
+@dataclass
+class Latent:
+    """
+    A latent extracted from a model's activations.
+    """
+
+    module_name: str
+    """The module name associated with the latent."""
+
+    latent_index: int
+    """The index of the latent within the module."""
+
+    def __repr__(self) -> str:
+        """
+        Return a string representation of the latent.
+
+        Returns:
+            str: A string representation of the latent.
+        """
+        return f"{self.module_name}_latent{self.latent_index}"
+
+
+class ActivationData(NamedTuple):
+    """
+    Represents the activation data for a latent.
+    """
+
+    locations: Float[Tensor, "n_examples 2"]
+    """Tensor of latent locations."""
+
+    activations: Float[Tensor, "n_examples"]
+    """Tensor of latent activations."""
+
+
+class LatentData(NamedTuple):
+    """
+    Represents the output of a TensorBuffer.
+    """
+
+    latent: Latent
+    """The latent associated with this output."""
+
+    module: str
+    """The module associated with this output."""
+
+    activation_data: ActivationData
+    """The activation data for this latent."""
 
 @dataclass
 class Neighbour:
@@ -61,28 +109,6 @@ class NonActivatingExample(Example):
     The distance from the neighbouring latent.
     Defaults to -1.0 if not using neighbours.
     """
-
-
-@dataclass
-class Latent:
-    """
-    A latent extracted from a model's activations.
-    """
-
-    module_name: str
-    """The module name associated with the latent."""
-
-    latent_index: int
-    """The index of the latent within the module."""
-
-    def __repr__(self) -> str:
-        """
-        Return a string representation of the latent.
-
-        Returns:
-            str: A string representation of the latent.
-        """
-        return f"{self.module_name}_latent{self.latent_index}"
 
 
 @dataclass
